fix: [2.6] remove golangci-lint v2 exclusion rules and fix lint violations (#49115)
## Summary
Cherry-pick of #48586 to the 2.6 branch.

- Remove all temporary exclusion rules added during the golangci-lint
v1→v2 upgrade
- Fix ~1500 lint violations (QF1008, ST1005, gosec, revive, ineffassign,
unconvert, depguard, gocritic, staticcheck)

pr: #48586

## Test plan
- [ ] CI passes


---------

+++ internal/datacoord/compaction_inspector.go
@@ -245,7 +245,7 @@ func (c *compactionInspector) schedule() []CompactionTask {
 
 	// The schedule loop will stop if either:
 	// 1. no more task to schedule (the task queue is empty)
-	// 2. no avaiable slots
+	// 2. no available slots
 	for {
 		t, err := c.queueTasks.Dequeue()
 		if err != nil {
@@ -408,7 +408,7 @@ func (c *compactionInspector) cleanCompactionTaskMeta() {
 		for _, task := range tasks {
 			if task.State == datapb.CompactionTaskState_cleaned {
 				duration := time.Since(time.Unix(task.StartTime, 0)).Seconds()
-				if duration > float64(Params.DataCoordCfg.CompactionDropToleranceInSeconds.GetAsDuration(time.Second).Seconds()) {
+				if duration > Params.DataCoordCfg.CompactionDropToleranceInSeconds.GetAsDuration(time.Second).Seconds() {
 					// try best to delete meta
 					err := c.meta.DropCompactionTask(context.TODO(), task)
 					log.Ctx(context.TODO()).Debug("drop compaction task meta", zap.Int64("planID", task.PlanID))
